Add rubocop (#373)
* adds rubocop with todo

* Layout, Style, Lint rules in rubocop.yml

* fixed: Layout/SpaceInsideHashLiteralBraces, UnusedBlockArguments,
TrailingWhiteSpace, TrailingBlankLines, UnusedMethodArguments

* updated rubocop by removing copies of blocks from the rubocop_todo.yml

* fixed: Style/AndOr, Style/BarePercentLiterals, Style/BracesAroundHashParameters, Style/BlockComments, Style/HashSyntax, Style/IfUnlessModifier

* fixed: Style/BlockDelimiters

+++ spec/models/dump_file_spec.rb
@@ -3,17 +3,15 @@ require 'fileutils'
 
 RSpec.describe DumpFile, type: :model do
 
-  subject { 
-    instance = described_class.create 
+  subject do
+    instance = described_class.create
     FileUtils.touch(instance.path)
     instance
-  }
+  end
 
-  after(:each) {
-    unless subject.nil? || !File.exists?(subject.path)
-      File.delete(subject.path)
-    end
-  }
+  after(:each) do
+    File.delete(subject.path) unless subject.nil? || !File.exists?(subject.path)
+  end
 
   describe 'when a new instance is created' do
     it 'populates :path' do
